Fix lack of thread joins causing deadlock and mapping leaks
A "hang" and resource leak were found during a stress test in which the
GetPathAction, ExePathAction and RecoveryAction servers were rapidly invoked
with many requests, some of which preempted the previous request.

For example, abstract_controller_execution::stopMoving() merely requests the
worker thread to exit but does not join() it. If ::startMoving() is called
shortly thereafter, the worker thread from the previous Action may not have
exited yet, in which case startMoving() does nothing since the thread is
"already running".  But that thread, for the previous Action exits thereafter,
leaving the abstract_controller_execution stuck waiting for the worker thread,
which isn't running anymore.

A solution would be to do thread_.join() in stopMoving() but this may delay
the result of the action server, especially if the reason is that patience
is exceeded. So a slightly better solution is to do the join() in
startMoving() just before starting the thread for the next Action, which is
how this code change fixes it.

A secondary effect of this bug, even when timing was such that it did not
"hang", was that the thread stack mappings are never freed due to lack of
::join()'s. After a few hours of stress testing, we had over 65,000 mappings
created for thread stacks at which point, we hit an OS limit, and the program
aborted with a boost::throw_exception<boost::thread_resource_error> from
boost::thread::start_thread. The addition of these joins solves this problem
also.

The same fix is applied to the abstract_planner_execution and
abstract_recovery_execution.

NOTE: I think that a better solution would be to create each type of worker
thread only once at start up and let it loop forever, processing each request
sequentially, as this would reduce some complexity and the overhead of
frequently creating and destroying threads, but this is a quicker fix for now.