Fixes lost controller thread
Because startMoving() would first try to join an old instance of the thread
(to avoid a thread leak), we could end up waiting in this join a long time if
the thread was already working on a goal.

Now we use a flag to track whether the thread is running or not and only
join() if it was running and we don't want it to be.