Makes canceling of already-running plugin thread-safe
When a new action goal is received on the same concurrency slot for which an
action was already running, the existing action was being canceled, but the
it did not wait for this to complete before starting a new one which, as noted
by a comment in the code, is not thread safe. Now it waits for the canceled
action thread to exit before starting a new one.

This lack of thread safety was causing multiple threads to call the same
plugin object concurrently, causing data corruption leading to a SIGSEGV.