RFC: Prevent actions from deeply queuing for a concurrency slot
While the previous commit prevents memory corruption caused by running a
plugin before the previous (canceled) action completes, it does not prevent
a waiting action from starting the plugin even though the client may have
already canceled.

As a result, if external code starts and cancels an action repeatedly in rapid
succession, all of them are run sequentially even though many of these actions
were already canceled.

This commit contains a partial fix:
If a new action is being started while one is already running in the same
concurrency slot, put it in a "pending" map indexed by slot, then tell the
existing plugin to cancel and wait for it to exit. After waiting, double check
that the current goal ID is still in the pending list before starting the
action plugin. If it's not, mark this goal as canceled and return.

Meanwhile, if the action client cancels a goal that is in the pending list,
remove it therefrom so that .start() will see that it is no longer pending as
described above and will not start the plugin.

FIXME:
While this does prevent the many actions that were already canceled by client code
from continuing to start long after they were canceled, it does so at the
expense of blocking the AbstractAction::start(), which is running on a "goal
callback" from an ActionServer. Since there are a limited number of threads
available for these and all ROS message callbacks, these can fill up, causing
errors such as the costmap not receiving sensor messages.

A proper fix will record pending actions to be processed by a worker queue and
will remove queued items when canceled without blocking the ActionServer
goal_cb.