mbf_costmap_nav: add 3D costmap navigation server
When configured to use 3D costmaps, instantiate a 3D costmap
navigation server. The only real difference between this and the
normal 2D costmap server is using the Costmap3DROS class, which
derives from the Costmap2DROS class. This allows planners at runtime
to detect that their 2D costmap is a 3D costmap as well and down-cast
to the 3D costmap to make 3D queries.