abstract controller: switch to steady clock
Priviously the thread clock was used to measure how much of the
control cycle was used. However, some controllers do all their work
in *other* threads defeating using the thread clock in this way. The
CPU clock is not a good guide either, as all threads will be lumped
into the clock. Just use the steady clock instead.