fix crashes on exit by making exit safe
The costmap nav server pointer was global and therefore could be
destroyed after certain global ROS objects (like the timer manager)
are already destroyed. This was causing crashes on exit. Fix this by
moving the pointer into the main function.

Also, ros::shutdown() was being used in the signal handler, which is
unsafe because it grabs a lock (the signal could be delivered while
the lock was held in the thread it was handled in). This was replaced
with the signal handler safe ros::requestShutdown().

There is still a mess with the plugins on exit where the plugins are
unloaded before the instances that were created from them have been
deleted. Pluginlib handles this OK by emitting a loud warning. The
plugin issue will be fixed in a later commit.