server node: use MultiThreadedSpinner
Add a new optional parameter, thread_count, to use as the number of
threads to pass to the new MultiThreadedSpinner. The thread_count
defaults to 0, which means one thread per core. A negative thread_count
means to go back to the old single-threaded spin(). A positive
thread_count means to spin a multi-threaded spinner with that number of
threads.